third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-pause/debugger-pause-on-failed-assertion-expected.txt

CONSOLE ERROR: line 9: console.assert
Tests that debugger breaks on failed assertion when pause on exception mode is enabled.

Script source was shown.
Script execution paused.
Call stack:
    0) failAssertion (debugger-pause-on-failed-assertion.html:9)
    1) handleClick (debugger-pause-on-failed-assertion.html:14)
Script execution resumed.
